Separate the default formatter into its own module
